+++ src/Interpreters/ThreadStatusExt.cpp
@@ -7,6 +7,7 @@
 #include <Interpreters/ProcessList.h>
 #include <Interpreters/QueryThreadLog.h>
 #include <Interpreters/QueryViewsLog.h>
+#include <Interpreters/TraceCollector.h>
 #include <Parsers/formatAST.h>
 #include <Common/CurrentThread.h>
 #include <Common/Exception.h>
@@ -14,7 +15,6 @@
 #include <Common/QueryProfiler.h>
 #include <Common/SensitiveDataMasker.h>
 #include <Common/ThreadProfileEvents.h>
-#include <Common/TraceCollector.h>
 #include <Common/setThreadName.h>
 #include <base/errnoToString.h>
